backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
pdfplumber>=0.10
pandas>=2.0
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
openpyxl>=3.1
